Make signature_test_inputs_different_msg generic
