+++ fastcrypto/src/tests/ed25519_tests.rs
@@ -182,26 +182,6 @@ fn signature_test_inputs() -> (Vec<u8>, Vec<Ed25519PublicKey>, Vec<Ed25519Signat
     (digest.to_vec(), pubkeys, signatures)
 }
 
-fn signature_test_inputs_different_msg(
-) -> (Vec<Vec<u8>>, Vec<Ed25519PublicKey>, Vec<Ed25519Signature>) {
-    // Make signatures.
-    let digests: Vec<Vec<u8>> = [b"Hello", b"world", b"!!!!!"]
-        .iter()
-        .map(|msg| hash::Sha256::digest(*msg).to_vec())
-        .collect();
-    let (pubkeys, signatures): (Vec<Ed25519PublicKey>, Vec<Ed25519Signature>) = keys()
-        .into_iter()
-        .take(3)
-        .zip(&digests)
-        .map(|(kp, digest)| {
-            let sig = kp.sign(digest.as_ref());
-            (kp.public().clone(), sig)
-        })
-        .unzip();
-
-    (digests, pubkeys, signatures)
-}
-
 #[test]
 fn verify_valid_batch() {
     let (digest, pubkeys, signatures) = signature_test_inputs();
@@ -247,14 +227,16 @@ fn verify_batch_missing_public_keys() {
 
 #[test]
 fn verify_valid_batch_different_msg() {
-    let (msgs, pks, sigs) = signature_test_inputs_different_msg();
+    let (msgs, pks, sigs) =
+        signature_tests::signature_test_inputs_different_msg::<Ed25519KeyPair>();
     let res = Ed25519PublicKey::verify_batch_empty_fail_different_msg(&msgs, &pks, &sigs);
     assert!(res.is_ok(), "{:?}", res);
 }
 
 #[test]
 fn verify_invalid_batch_different_msg() {
-    let (msgs, pks, mut sigs) = signature_test_inputs_different_msg();
+    let (msgs, pks, mut sigs) =
+        signature_tests::signature_test_inputs_different_msg::<Ed25519KeyPair>();
     sigs[0] = Ed25519Signature::default();
     let res = Ed25519PublicKey::verify_batch_empty_fail_different_msg(&msgs, &pks, &sigs);
     assert!(res.is_err(), "{:?}", res);
